fix: converting primitive types to reference types

+++ src/main/java/ru/yandex/practicum/filmorate/dao/film/FilmStorage.java
@@ -4,7 +4,7 @@ import ru.yandex.practicum.filmorate.model.*;
 import java.util.Collection;
 
 public interface FilmStorage {
-    Film getFilmsById(Integer id);
+    Film getFilmsById(Long id);
 
     Collection<Film> getAllFilms();
 
@@ -22,11 +22,11 @@ public interface FilmStorage {
 
     Film updateFilm(Film film);
 
-    void deleteFilm(Integer id);
+    void deleteFilm(Long id);
 
-    Film putLike(Integer id, Long userId);
+    Film putLike(Long id, Long userId);
 
-    Film deleteLike(Integer id, Long userId);
+    Film deleteLike(Long id, Long userId);
 
-    boolean isExistsIdFilm(Integer filmId);
+    boolean isExistsIdFilm(Long filmId);
 }
