+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -11,7 +11,7 @@ import java.util.*;
 @RequiredArgsConstructor
 @AllArgsConstructor
 public class Film {
-    private int id;
+    private Long id;
     @NotBlank
     private String name;
     @NotBlank
@@ -23,12 +23,12 @@ public class Film {
     private LocalDate releaseDate;
     @NotNull
     @Positive
-    private int duration;
+    private Integer duration;
     @NotNull
     private RatingMpa mpa;
     private List<Genre> genres = new ArrayList<>();
 
-    public Film(String name, String description, LocalDate releaseDate, int duration,
+    public Film(String name, String description, LocalDate releaseDate, Integer duration,
                 RatingMpa mpa, List<Genre> genres) {
         this.name = name;
         this.description = description;
